Time: 0 ms (100%), Space: 9.2 MB (51.16%) - LeetHub

+++ 0389-find-the-difference/0389-find-the-difference.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     char findTheDifference(string s, string t) {
         for (char c : t) {
-            size_t pos = s.find(c);
+            int pos = s.find(c);
             if (pos != string::npos)
                 s.erase(pos, 1);
             else
